requirements.txt
absl-py==0.7.0
bleach==1.5.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
Flask==1.0.2
h5py==2.9.0
html5lib==0.9999999
idna==2.8
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
MarkupSafe==1.1.1
numpy==1.16.2
opencv-contrib-python==3.4.4.19
opencv-python==3.4.4.19
python-dotenv==0.9.1
Pillow==5.4.1
protobuf==3.7.0
PyYAML==3.13
requests==2.21.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
urllib3==1.24.1
Werkzeug==0.14.1
